fix: Correct test range completion logic
When running the Structured - Test Protocol Mid test workflow, the phone begins the 6ft 0° test at the end of the first "station" resulting in that data being captured in the log file. This was as a result of the logic using the number of locations, rather than the number of angles when determining if the test range had been completed or not. All of the other test workflows perform correctly because the number of angles and locations are the same.